Major refactoring for cuda backend pooling
Following #5198 moves pooling functionality from blas to separate file
`pool.py`. Similarly, tests moved to separate file. New addition:
average grad pooling (which was not present in blas before) and merge of
maxpool and average pool in one class GpuPool.
