+++ theano/sandbox/cuda/dnn.py
@@ -27,7 +27,8 @@ from theano.sandbox.cuda.basic_ops import (as_cuda_ndarray_variable,
                                            gpu_alloc, GpuAlloc,
                                            gpu_alloc_empty, GpuAllocEmpty,
                                            GpuElemwise)
-from theano.sandbox.cuda.blas import GpuConv, GpuMaxPool, GpuMaxPoolGrad
+from theano.sandbox.cuda.blas import GpuConv
+from theano.sandbox.cuda.pool import GpuPool, GpuMaxPoolGrad, GpuAveragePoolGrad
 from theano.sandbox.cuda.nnet import GpuSoftmax
 from theano.sandbox.cuda.opt_util import (alpha_merge, output_merge,
                                           pad_dims, unpad_dims)
@@ -3037,11 +3038,11 @@ if True:
         return [node.op(*inputs)]
 
     @register_opt('cudnn')
-    @local_optimizer([GpuMaxPool])
+    @local_optimizer([GpuPool])
     def local_pool_dnn(node):
         if not dnn_available():
             return
-        if isinstance(node.op, GpuMaxPool):
+        if isinstance(node.op, GpuPool):
             if not node.op.ignore_border:
                 return
             img, = node.inputs
@@ -3076,20 +3077,19 @@ if True:
                 return [host_from_gpu(ret)]
 
     @register_opt('cudnn')
-    @local_optimizer([GpuMaxPoolGrad])
+    @local_optimizer([GpuMaxPoolGrad, GpuAveragePoolGrad])
     def local_pool_dnn_grad(node):
         if not dnn_available():
             return
-        if isinstance(node.op, GpuMaxPoolGrad):
+        if isinstance(node.op, (GpuMaxPoolGrad, GpuAveragePoolGrad)):
             if not node.op.ignore_border:
                 return
-            inp, out, inp_grad = node.inputs
-            ds = node.op.ds
+            inp, out, inp_grad, ws, stride, pad = node.inputs
 
-            return [GpuDnnPoolGrad(mode='max')(gpu_contiguous(inp),
-                                               gpu_contiguous(out),
-                                               gpu_contiguous(inp_grad),
-                                               ds, ds, (0, 0))]
+            return [GpuDnnPoolGrad(mode=node.op.mode)(gpu_contiguous(inp),
+                                                      gpu_contiguous(out),
+                                                      gpu_contiguous(inp_grad),
+                                                      ws, stride, pad)]
 
     @register_opt('cudnn')
     @local_optimizer([MaxPoolGrad])
